itest: extract channel setup, move helper functions
This commit extracts the creation of the asset test channel network into
its own function. At the same time we also move all helper functions to
a new file.
